bulid:网关集成URL资源权限校验

+++ basic-uums/basic-uums-entity/src/main/java/com/basic/cloud/uums/entity/ResourceAuthority.java
@@ -37,7 +37,7 @@ public class ResourceAuthority extends BisDataEntity<ResourceAuthority> {
     /**
      * 资源操作类型,1可以使用,2可分配
      */
-    private Long resourceOpt;
+    private Integer resourceOpt;
 
     /**
      * 描述
